refactor(mb-api): extract computeTauxProgressionPoints pipeline helper
Suite review PR #2210 : le pipeline (load contexts + resolveSerieIndividu +
resolveObjectifIndividu + resolveTauxProgression) était dupliqué entre
listTauxProgressionForIndicateur (1 indicateur, N cibles) et computeDernierTaux
côté panier (1 indicateur, 1 cible). Extraction d'un helper unique dans
valeurAvancement/queries/. Les deux queries n'en gardent que leur logique
propre : filtres dateDebut/dateFin + tri pour la liste, sélection du max
bucket pour le panier.
