+++ apps/mb-api/src/valeurAvancement/queries/listTauxProgressionForIndicateur.ts
@@ -7,27 +7,12 @@ import {
 import { ResultAsync } from 'neverthrow'
 
 import { requireCurrentPrincipalId } from '@/framework/auth/userContext'
-import { type BucketKey, compareBuckets, formatBucket, parseBucket } from '@/framework/bucket'
+import { compareBuckets, formatBucket, parseBucket } from '@/framework/bucket'
 import { logger } from '@/framework/logger/logger'
 import { db } from '@/framework/persistence/dbStore'
-import { loadIndividusParPublicId } from '@/indicateur/queries/loadIndicateurIndividuContext'
 import { withIndicateurReadPermission } from '@/indicateur/permissions'
-import { loadResolveObjectifContext } from '@/objectifIndicateurIndividu/queries/loadResolveObjectifContext'
-import {
-  type PointObjectifInterne,
-  resolveObjectifIndividu,
-} from '@/objectifIndicateurIndividu/resolveObjectifIndividu'
-import {
-  type ObjectifBrut,
-  resolveTauxProgression,
-  type ValeurBrute,
-} from '@/valeurAvancement/resolveTauxProgression'
-import { loadResolveSerieContext } from '@/valeurAvancement/queries/loadResolveSerieContext'
-import {
-  type IndividuRef,
-  type PointInterne,
-  resolveSerieIndividu,
-} from '@/valeurAvancement/resolveSerieIndividu'
+import { loadIndividusParPublicId } from '@/indicateur/queries/loadIndicateurIndividuContext'
+import { computeTauxProgressionPoints } from '@/valeurAvancement/queries/computeTauxProgressionPoints'
 
 // Défauts compromis lisibilité × coût : mensuel des deux côtés (cf. doc archi
 // `taux-progression.md`). Surchargeable par requête, avec la contrainte
@@ -62,38 +47,13 @@ const buildList = async ({
   const dateTruncValeur = params.dateTruncValeur ?? DEFAULT_DATE_TRUNC_VALEUR
   const dateTruncObjectif = params.dateTruncObjectif ?? DEFAULT_DATE_TRUNC_OBJECTIF
   const startedAt = performance.now()
-  const [{ ctx: serieCtx, allNodes }, { ctx: objectifCtx }] = await Promise.all([
-    loadResolveSerieContext({
-      indicateurId: indicateur.id,
-      cibles: individusCibles,
-      dateTrunc: dateTruncValeur,
-    }),
-    loadResolveObjectifContext({
-      indicateurId: indicateur.id,
-      cibles: individusCibles,
-      dateTrunc: dateTruncObjectif,
-    }),
-  ])
-  const serieCache = new Map<string, ReadonlyArray<PointInterne>>()
-  const objectifCache = new Map<string, ReadonlyMap<BucketKey, PointObjectifInterne>>()
 
-  const valeurs: ValeurBrute[] = []
-  const objectifsParIndividu = new Map<string, ObjectifBrut[]>()
-
-  for (const cible of individusCibles) {
-    const serie = await resolveSerieIndividu(cible.id, serieCtx, serieCache)
-    for (const point of serie) {
-      valeurs.push(toValeurBrute({ cible, point }))
-    }
-
-    const objectifsMap = resolveObjectifIndividu(cible.id, objectifCtx, objectifCache)
-    const objectifsList = toObjectifBruts(objectifsMap)
-    if (objectifsList.length > 0) {
-      objectifsParIndividu.set(cible.id, objectifsList)
-    }
-  }
-
-  const points = resolveTauxProgression({ valeurs, objectifsParIndividu })
+  const points = await computeTauxProgressionPoints({
+    indicateurId: indicateur.id,
+    cibles: individusCibles,
+    dateTruncValeur,
+    dateTruncObjectif,
+  })
 
   // Filtres dateDebut/dateFin appliqués en sortie pour ne pas perturber le
   // carry-forward des séries dérivées (cf. design doc indicateur-derives.md).
@@ -128,7 +88,6 @@ const buildList = async ({
       dateTruncValeur,
       dateTruncObjectif,
       nbCibles: individusCibles.length,
-      nbNodes: allNodes.size,
       nbPoints: items.length,
       durationMs: Math.round(performance.now() - startedAt),
     },
@@ -136,23 +95,3 @@ const buildList = async ({
   )
   return { items }
 }
-
-const toValeurBrute = ({
-  cible,
-  point,
-}: {
-  cible: IndividuRef
-  point: PointInterne
-}): ValeurBrute => ({
-  individuId: cible.id,
-  individuPublicId: cible.publicId,
-  date: point.bucket,
-  valeur: point.valeur,
-})
-
-const toObjectifBruts = (
-  objectifsMap: ReadonlyMap<BucketKey, PointObjectifInterne>,
-): ObjectifBrut[] =>
-  [...objectifsMap.values()]
-    .map((point) => ({ dateCible: point.bucket, valeurCible: point.valeur }))
-    .sort((a, b) => compareBuckets(a.dateCible, b.dateCible))
